requirements.txt
mysql-connector-python
nltk
numpy
gensim
ijson
graphviz
pandas
